Feat: Add QuizModal component
